Fix seek to accept decimal values for precise seeking
- Changed iOS seek to use getDouble instead of getInt
- Changed Android seek to accept double instead of long
- Fixes issue with loop functionality where precise seek positions are required

+++ android/src/main/java/us/mediagrid/capacitorjs/plugins/nativeaudio/AudioSource.java
@@ -147,8 +147,8 @@ public class AudioSource extends Binder {
         getPlayer().pause();
     }
 
-    public void seek(long timeInSeconds) {
-        getPlayer().seekTo(timeInSeconds * 1000);
+    public void seek(double timeInSeconds) {
+        getPlayer().seekTo((long) (timeInSeconds * 1000));
     }
 
     public void stop() {
